introduce getDynamicSamplingContextFromClient
